--- apps/webapp/components/icons/blog/index.ts
@@ -1,3 +1,0 @@
-export * from "./logo-main.icon"
-export * from "./logo-blog.icon"
-export * from "./lucide-react.icon"
