--- apps/webapp/components/icons/blog/logo-blog.icon.tsx
@@ -1,25 +1,0 @@
-import { cn } from "@/lib/utils.lib"
-
-export function LogoBlog(
-  props: React.ComponentPropsWithoutRef<"svg"> & { blog?: boolean }
-) {
-  return (
-    <svg width="48" height="48" viewBox="0 0 48 48" fill="none" {...props}>
-      <g clipPath="url(#clip0_97_1605)">
-        <path
-          d="M23.958 47.916C37.1896 47.916 47.916 37.1896 47.916 23.958C47.916 10.7264 37.1896 0 23.958 0C10.7264 0 0 10.7264 0 23.958C0 37.1896 10.7264 47.916 23.958 47.916Z"
-          className="fill-primary-200  dark:fill-white"
-        />
-        <path
-          d="M32.736 27.642C32.7752 28.6237 32.5666 29.5994 32.1293 30.4792C31.692 31.359 31.0402 32.1145 30.234 32.676C29.1703 33.4241 27.9628 33.943 26.688 34.2C26.6938 34.2699 26.6938 34.3401 26.688 34.41C26.653 35.2005 26.3144 35.947 25.7427 36.494C25.171 37.0409 24.4102 37.3463 23.619 37.3463C22.8278 37.3463 22.067 37.0409 21.4953 36.494C20.9236 35.947 20.585 35.2005 20.55 34.41C20.5442 34.3081 20.5442 34.2059 20.55 34.104C18.5633 33.6461 16.7131 32.725 15.15 31.416L17.85 27.66C19.4485 29.0839 21.5173 29.8661 23.658 29.856C24.4589 29.8905 25.2576 29.7492 25.998 29.442C26.2474 29.3519 26.4644 29.1897 26.6214 28.976C26.7784 28.7624 26.8684 28.5069 26.88 28.242C26.8879 27.9848 26.8027 27.7334 26.64 27.534C26.4193 27.2809 26.1358 27.0905 25.818 26.982C25.4675 26.8235 25.1042 26.6951 24.732 26.598C24.402 26.502 23.91 26.376 23.274 26.232L22.2 25.938C21.93 25.878 21.564 25.776 21.084 25.644C20.604 25.512 20.214 25.374 19.932 25.26C19.65 25.146 19.29 24.978 18.888 24.786C18.5252 24.6223 18.1862 24.4103 17.88 24.156C17.591 23.9229 17.3184 23.6704 17.064 23.4C16.7864 23.113 16.5574 22.7827 16.386 22.422C16.2139 22.0358 16.0793 21.634 15.984 21.222C15.8732 20.744 15.8188 20.2547 15.822 19.764C15.7928 18.468 16.2007 17.1999 16.98 16.164C17.7577 15.1531 18.8094 14.3867 20.01 13.956C20.1922 13.882 20.381 13.8257 20.574 13.788C20.5682 13.7181 20.5682 13.6479 20.574 13.578C20.5933 12.7853 20.9184 12.0307 21.4813 11.4722C22.0441 10.9136 22.8012 10.5944 23.594 10.5812C24.3869 10.568 25.1542 10.8619 25.7353 11.4015C26.3164 11.941 26.6664 12.6844 26.712 13.476C28.7597 13.8388 30.6935 14.6773 32.358 15.924L29.628 19.674C28.1323 18.5567 26.3282 17.9281 24.462 17.874C23.7706 17.8487 23.0816 17.9693 22.44 18.228C22.2107 18.3094 22.0117 18.4586 21.8692 18.6558C21.7267 18.853 21.6474 19.0888 21.642 19.332C21.6433 19.5478 21.7125 19.7578 21.84 19.932C21.9646 20.1078 22.1214 20.2585 22.302 20.376C22.5561 20.5274 22.8283 20.6464 23.112 20.73C23.4259 20.8438 23.7464 20.938 24.072 21.012C24.354 21.09 24.756 21.174 25.272 21.3L26.514 21.564C26.832 21.642 27.252 21.75 27.762 21.9C28.2028 22.022 28.634 22.1764 29.052 22.362L30.198 22.962C30.5708 23.1629 30.9091 23.4221 31.2 23.73C31.4894 24.0416 31.7523 24.3767 31.986 24.732C32.2513 25.1269 32.4386 25.5688 32.538 26.034C32.6708 26.5597 32.7373 27.0998 32.736 27.642Z"
-          className="fill-white  dark:fill-primary-200"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_97_1605">
-          <rect width="48" height="48" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  )
-}
