--- apps/webapp/components/icons/blog/logo-main.icon.tsx
@@ -1,83 +1,0 @@
-export function LogoMain(
-  props: React.ComponentPropsWithoutRef<"svg"> & { blog?: boolean }
-) {
-  const { blog, ...rest } = props
-  const blogValue = blog
-
-  return (
-    <svg viewBox="0 0 317 74" fill="none" {...rest}>
-      <path
-        d="M121.025 27.0562C123.073 27.0095 125.109 27.4113 126.989 28.2339C128.868 29.0563 130.548 30.2799 131.912 31.82C134.834 35.0628 136.392 39.3207 136.258 43.697C136.395 48.0765 134.837 52.3382 131.912 55.5832C130.547 57.1214 128.867 58.3436 126.987 59.166C125.108 59.9883 123.073 60.3912 121.025 60.347C119.255 60.3921 117.495 60.0668 115.856 59.3917C114.218 58.7166 112.736 57.7064 111.506 56.425V59.5885H102.465V14.6982H111.506V30.9782C112.735 29.6951 114.216 28.6836 115.855 28.0083C117.494 27.3331 119.254 27.0089 121.025 27.0562ZM113.674 49.2562C115.157 50.6877 117.133 51.4869 119.187 51.4869C121.241 51.4869 123.217 50.6877 124.7 49.2562C126.116 47.7398 126.904 45.7367 126.904 43.6554C126.904 41.574 126.116 39.5709 124.7 38.0545C123.217 36.6231 121.241 35.8239 119.187 35.8239C117.133 35.8239 115.157 36.6231 113.674 38.0545C112.258 39.5709 111.47 41.574 111.47 43.6554C111.47 45.7367 112.258 47.7398 113.674 49.2562Z"
-        className={
-          blogValue
-            ? "dark:fill-black fill-white "
-            : "dark:fill-white  fill-black"
-        }
-      />
-      <path
-        d="M146.209 23.7355C145.475 23.7506 144.744 23.6162 144.063 23.3404C143.381 23.0646 142.762 22.653 142.242 22.1305C141.721 21.608 141.312 20.9851 141.037 20.2993C140.762 19.6134 140.627 18.8788 140.641 18.1392C140.621 17.4031 140.754 16.6708 141.031 15.9892C141.307 15.3075 141.722 14.6915 142.249 14.1802C142.769 13.6614 143.388 13.2538 144.068 12.9821C144.749 12.7102 145.477 12.5797 146.209 12.5985C146.952 12.5745 147.693 12.7023 148.386 12.9741C149.079 13.246 149.71 13.6562 150.242 14.1802C150.77 14.691 151.186 15.3068 151.464 15.9884C151.743 16.67 151.878 17.4025 151.86 18.1392C151.879 18.8816 151.746 19.62 151.467 20.3078C151.189 20.9955 150.772 21.6178 150.242 22.1352C149.712 22.6626 149.082 23.0762 148.388 23.3513C147.696 23.6263 146.954 23.757 146.209 23.7355ZM141.707 59.5885V27.8147H150.739V59.5885H141.707Z"
-        className={
-          blogValue
-            ? "dark:fill-black fill-white "
-            : "dark:fill-white  fill-black"
-        }
-      />
-      <path
-        d="M177.825 36.075H169.721V47.73C169.721 50.2645 171.041 51.5318 173.681 51.5318C175.133 51.4982 176.554 51.1109 177.825 50.4033V58.9595C175.959 59.894 173.9 60.3694 171.816 60.347C168.104 60.347 165.329 59.348 163.492 57.35C161.654 55.352 160.735 52.648 160.735 49.2378V36.075H155.957V27.75H160.735V19.9338H169.767V27.8148H177.871L177.825 36.075Z"
-        className={
-          blogValue
-            ? "dark:fill-black fill-white "
-            : "dark:fill-white  fill-black"
-        }
-      />
-      <path
-        d="M197.745 60.347C195.505 60.4091 193.275 60.018 191.189 59.1965C189.103 58.375 187.201 57.14 185.598 55.5647C184.008 54.038 182.752 52.1939 181.91 50.1503C181.068 48.1067 180.659 45.9088 180.71 43.697C180.659 41.4853 181.068 39.2875 181.911 37.244C182.752 35.2005 184.008 33.3562 185.598 31.8292C187.205 30.2547 189.111 29.0209 191.2 28.201C193.289 27.3812 195.521 26.9919 197.763 27.0562C200.746 27.017 203.689 27.7428 206.317 29.1652C208.812 30.517 210.901 32.5204 212.363 34.965L204.489 39.775C203.817 38.5728 202.837 37.5745 201.649 36.8845C200.463 36.1945 199.114 35.8383 197.745 35.853C196.722 35.827 195.705 36.0132 194.756 36.3998C193.807 36.7866 192.948 37.3655 192.232 38.1007C190.816 39.6172 190.027 41.6203 190.027 43.7016C190.027 45.783 190.816 47.786 192.232 49.3025C192.948 50.0377 193.807 50.6166 194.756 51.0034C195.705 51.39 196.722 51.5761 197.745 51.5502C199.114 51.5649 200.463 51.2087 201.649 50.5187C202.837 49.8287 203.817 48.8305 204.489 47.6282L212.344 52.4012C210.893 54.8613 208.803 56.8761 206.299 58.2287C203.674 59.6593 200.729 60.3887 197.745 60.347Z"
-        className={
-          blogValue
-            ? "dark:fill-black fill-white "
-            : "dark:fill-white  fill-black"
-        }
-      />
-      <path
-        d="M239.219 27.8148H248.251V59.5885H239.219V56.425C237.986 57.7028 236.504 58.7106 234.866 59.3855C233.229 60.0604 231.47 60.3876 229.7 60.347C227.653 60.3886 225.621 59.9828 223.744 59.1571C221.868 58.3315 220.192 57.1058 218.831 55.5648C215.9 52.3298 214.335 48.0735 214.467 43.697C214.336 39.3207 215.901 35.0647 218.831 31.8293C220.194 30.2914 221.87 29.0684 223.747 28.2445C225.623 27.4206 227.654 27.0153 229.7 27.0563C231.476 27.0071 233.242 27.3304 234.888 28.0056C236.533 28.6808 238.02 29.6931 239.256 30.9783L239.219 27.8148ZM231.474 51.5503C232.496 51.5751 233.513 51.3884 234.462 51.0018C235.41 50.6152 236.269 50.0369 236.987 49.3025C238.402 47.7861 239.191 45.783 239.191 43.7016C239.191 41.6203 238.402 39.6172 236.987 38.1008C236.269 37.3664 235.41 36.7881 234.462 36.4015C233.513 36.0148 232.496 35.8281 231.474 35.853C230.45 35.8203 229.43 36.0033 228.481 36.3904C227.532 36.7776 226.674 37.3601 225.961 38.1008C224.555 39.6224 223.773 41.6235 223.773 43.7016C223.773 45.7798 224.555 47.7809 225.961 49.3025C226.674 50.0432 227.532 50.6257 228.481 51.0128C229.43 51.3999 230.45 51.5829 231.474 51.5503Z"
-        className={
-          blogValue
-            ? "dark:fill-black fill-white "
-            : "dark:fill-white  fill-black"
-        }
-      />
-      <path
-        d="M266.544 60.347C261.599 60.4167 256.793 58.6966 253.001 55.5L257.2 49.6263C259.694 51.8485 262.911 53.0727 266.241 53.0673C267.5 53.1232 268.756 52.8957 269.916 52.4013C270.302 52.2646 270.639 52.016 270.884 51.6867C271.129 51.3574 271.273 50.9622 271.295 50.5513C271.295 50.1436 271.163 49.7469 270.918 49.4228C270.566 49.0319 270.124 48.7335 269.632 48.5533C269.084 48.321 268.523 48.1232 267.95 47.9613L265.672 47.3785L263.935 46.99C263.519 46.8913 262.94 46.7341 262.198 46.5183C261.595 46.3516 260.999 46.154 260.416 45.9263C259.966 45.7413 259.424 45.4915 258.78 45.177C258.223 44.9174 257.699 44.5913 257.219 44.2058C256.833 43.8635 256.41 43.4565 255.951 42.9663C255.522 42.5168 255.171 41.9977 254.912 41.4308C254.646 40.8355 254.437 40.216 254.288 39.5808C254.115 38.8317 254.028 38.0651 254.03 37.296C253.99 35.2514 254.638 33.253 255.868 31.6258C257.081 30.045 258.715 28.845 260.581 28.1663C262.701 27.4082 264.938 27.0385 267.188 27.0748C271.72 27.0399 276.138 28.502 279.766 31.2373L275.457 37.0925C273.118 35.369 270.314 34.4012 267.417 34.3175C266.336 34.2771 265.258 34.4695 264.257 34.8818C263.903 35.0084 263.596 35.2412 263.377 35.5489C263.159 35.8564 263.039 36.2241 263.035 36.6023C263.034 36.9354 263.14 37.2599 263.338 37.5273C263.533 37.8067 263.779 38.0454 264.064 38.2303C264.458 38.4648 264.877 38.6541 265.313 38.7945C265.805 38.9688 266.305 39.117 266.811 39.2385L268.648 39.664L270.587 40.0988C271.065 40.219 271.717 40.3855 272.535 40.6168C273.218 40.8063 273.885 41.0507 274.529 41.3475C275.043 41.6065 275.631 41.9118 276.293 42.2725C276.906 42.59 277.46 43.0125 277.928 43.5213C278.375 44.0166 278.784 44.5453 279.15 45.103C279.558 45.723 279.851 46.4125 280.014 47.138C280.204 47.9659 280.297 48.8135 280.29 49.6633C280.35 51.1984 280.026 52.7242 279.347 54.0999C278.668 55.4756 277.655 56.657 276.403 57.535C273.495 59.4994 270.043 60.484 266.544 60.347Z"
-        className={
-          blogValue
-            ? "dark:fill-black fill-white "
-            : "dark:fill-white  fill-black"
-        }
-      />
-      <path
-        d="M303.811 27.0562C305.418 26.9843 307.022 27.2397 308.528 27.8071C310.035 28.3746 311.411 29.2424 312.576 30.3585C313.699 31.5138 314.574 32.8894 315.146 34.3995C315.718 35.9097 315.976 37.5222 315.902 39.1367V59.5885H306.88V42.2077C306.921 41.3643 306.795 40.5212 306.51 39.7273C306.224 38.9332 305.785 38.2043 305.217 37.5827C304.643 36.9998 303.953 36.5446 303.194 36.2466C302.435 35.9486 301.621 35.8145 300.807 35.853C299.938 35.8148 299.072 35.9522 298.257 36.2573C297.443 36.5624 296.698 37.0288 296.066 37.629C295.452 38.2406 294.971 38.9743 294.654 39.7834C294.338 40.5925 294.193 41.4591 294.228 42.328V59.5885H285.196V14.6982H294.228V31.1632C295.441 29.8285 296.924 28.7711 298.577 28.0626C300.23 27.3541 302.015 27.0109 303.811 27.0562Z"
-        className={
-          blogValue
-            ? "dark:fill-black fill-white "
-            : "dark:fill-white  fill-black"
-        }
-      />
-      <path
-        d="M36.6878 73.8705C56.9498 73.8705 73.3755 57.334 73.3755 36.9352C73.3755 16.5365 56.9498 0 36.6878 0C16.4257 0 0 16.5365 0 36.9352C0 57.334 16.4257 73.8705 36.6878 73.8705Z"
-        className={
-          blogValue
-            ? "dark:fill-primary-200 fill-white "
-            : "dark:fill-white  fill-primary-200"
-        }
-      />
-      <path
-        d="M50.1298 42.6148C50.1899 44.1282 49.8704 45.6325 49.2007 46.9888C48.5311 48.3451 47.533 49.5099 46.2984 50.3755C44.6695 51.5288 42.8204 52.3288 40.8683 52.725C40.8772 52.8328 40.8772 52.941 40.8683 53.0488C40.8147 54.2675 40.2962 55.4182 39.4207 56.2616C38.5452 57.1048 37.3803 57.5755 36.1686 57.5755C34.957 57.5755 33.7921 57.1048 32.9165 56.2616C32.041 55.4182 31.5225 54.2675 31.469 53.0488C31.4601 52.8916 31.4601 52.7342 31.469 52.577C28.4267 51.8711 25.5933 50.4511 23.1997 48.433L27.3343 42.6425C29.7822 44.8377 32.9502 46.0436 36.2283 46.028C37.4548 46.0812 38.6779 45.8634 39.8117 45.3898C40.1935 45.2508 40.5259 45.0008 40.7663 44.6714C41.0068 44.342 41.1446 43.9482 41.1623 43.5398C41.1744 43.1432 41.0439 42.7556 40.7948 42.4483C40.4569 42.058 40.0227 41.7645 39.536 41.5973C38.9993 41.353 38.4429 41.1549 37.873 41.0053C37.3677 40.8573 36.6142 40.663 35.6403 40.441L33.9957 39.9878C33.5822 39.8953 33.0217 39.738 32.2867 39.5345C31.5516 39.331 30.9544 39.1183 30.5226 38.9425C30.0908 38.7668 29.5395 38.5078 28.9239 38.2118C28.3684 37.9593 27.8491 37.6326 27.3803 37.2405C26.9378 36.8812 26.5203 36.4918 26.1307 36.075C25.7056 35.6325 25.355 35.1233 25.0925 34.5673C24.8289 33.9718 24.6228 33.3524 24.4769 32.7173C24.3072 31.9803 24.2239 31.226 24.2288 30.4695C24.1841 28.4715 24.8087 26.5165 26.0021 24.9195C27.193 23.361 28.8035 22.1795 30.642 21.5155C30.921 21.4014 31.2102 21.3146 31.5057 21.2565C31.4968 21.1487 31.4968 21.0405 31.5057 20.9328C31.5353 19.7106 32.0331 18.5474 32.895 17.6863C33.757 16.8252 34.9163 16.333 36.1304 16.3127C37.3445 16.2924 38.5195 16.7455 39.4093 17.5773C40.2992 18.4091 40.8352 19.5551 40.9051 20.7755C44.0407 21.3349 47.0021 22.6275 49.551 24.5495L45.3704 30.3308C43.0799 28.6083 40.3173 27.6392 37.4595 27.5558C36.4008 27.5167 35.3457 27.7027 34.3632 28.1015C34.0121 28.227 33.7072 28.4571 33.489 28.761C33.2708 29.065 33.1494 29.4285 33.1412 29.8035C33.1431 30.1362 33.2491 30.46 33.4444 30.7285C33.6352 30.9995 33.8753 31.2318 34.1519 31.413C34.541 31.6465 34.9577 31.8298 35.3922 31.9588C35.8729 32.1342 36.3638 32.2795 36.8623 32.3935C37.2942 32.5138 37.9098 32.6433 38.6999 32.8375L40.6018 33.2445C41.0888 33.3648 41.732 33.5313 42.513 33.7625C43.1879 33.9507 43.8482 34.1887 44.4884 34.4748L46.2433 35.3998C46.8142 35.7095 47.3323 36.1091 47.7777 36.5838C48.2208 37.0641 48.6234 37.5808 48.9813 38.1285C49.3876 38.7373 49.6745 39.4185 49.8266 40.1358C50.03 40.9462 50.1318 41.7789 50.1298 42.6148Z"
-        className={
-          blogValue
-            ? "dark:fill-white fill-primary-200 "
-            : "dark:fill-primary-200  fill-white"
-        }
-      />
-    </svg>
-  )
-}
