+++ apps/webapp/components/routes/blog/tag.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils.lib"
+import { cn } from '@/lib/utils'
 
 export const Tag = (props: any) => {
   const { title } = props
@@ -6,12 +6,12 @@ export const Tag = (props: any) => {
     <button
       {...props}
       className={cn(
-        "duration-400 disabled:opacity-2 group flex w-max items-center rounded-full bg-black px-space-12 py-space-6  leading-none text-primary-200 transition ease-in-out focus-within:gap-3 hover:-translate-y-0.5 hover:scale-105 hover:gap-3 hover:bg-neutral-900 focus:text-neutral-200 disabled:cursor-not-allowed disabled:text-neutral-200 dark:bg-white dark:text-primary-800 dark:hover:bg-neutral-200 dark:focus:text-neutral-700",
+        'duration-400 disabled:opacity-2 px-space-12 py-space-6 text-primary-200 dark:text-primary-800 group flex w-max items-center  rounded-full bg-black leading-none transition ease-in-out focus-within:gap-3 hover:-translate-y-0.5 hover:scale-105 hover:gap-3 hover:bg-neutral-900 focus:text-neutral-200 disabled:cursor-not-allowed disabled:text-neutral-200 dark:bg-white dark:hover:bg-neutral-200 dark:focus:text-neutral-700',
         props.className,
         props.classNameTitle
       )}
     >
-      {title.replace("_", " ")}
+      {title.replace('_', ' ')}
 
       {/* // * Temporally disabled. No UX currently using close property. */}
       {/* {props.close && (
