+++ apps/webapp/services/datocms/index.ts
@@ -1,4 +1,5 @@
-export * from "./datocms-layout.service"
-export * from "./datacms-blog-category.service"
-export * from "./datocms-seo.service"
-export * from "./datocms-page.service"
+export * from './datocms-layout.service'
+export * from './datacms-blog-category.service'
+export * from './datocms-seo.service'
+export * from './datocms-page.service'
+export * from './datocms-blog.service'
